WIP
The server integration with the VS extension seems to be connecting, but it never hits the DidTextChange. I'm missing something, but I'm not sure what it is, as I did have it trigger one time.  Once the didtextchange is triggered, the rest I think should work to be uncommneted to run the analysis and push diagnostics.