Append Job Id and Attempt to Pipeline Signed Artifacts
If the pipeline fails during deployment the failed job cannot be rerun because this task will fail because the signed_binaries artifact already exists. However, the entire task also can't be re-run because the extensions have already been published by the time it hits this point. This PR adds a unique value so the failed job can be rerun in the case of temporary failures like API key rotations.