Remove VS Code plugin build step from pipelines (#699)
* Remove VS Code plugin build step from pipelines

Eliminated the Npm build task for the VS Code plugin from both PR and release pipeline YAML files - the pipeline-pack step should be sufficient to create the extension file and performs the required steps to build properly in pipeline specficially.

* Update changelog for pipeline build fix

Added entry for version 1.0.63 on 2025-07-30 documenting the fix for VS Code Pipeline Build.